[pydrake] Merge all visualization API into the top-level module
Our design intention for the pydrake API moving forward is a single
package "pydrake" with one level of modules (e.g., "visualization")
under it.

The main programs like pydrake.visualization.meldis can remain as
additionally-nested modules, but for users' imports we want to provide
shorter paths with just a single dot in the module name.

Deprecate the older (longer) paths.

Fix some spelling typos that were highlighted by the move.
