+++ bindings/pydrake/visualization/BUILD.bazel
@@ -46,9 +46,11 @@ drake_pybind_library(
         "//lcmtypes:lcmtypes_drake_py",
     ],
     py_srcs = [
-        "all.py",
         "meldis.py",
         "plotting.py",
+        "_meldis.py",
+        "_plotting.py",
+        "_visualization_extra.py",
     ],
 )
 
@@ -86,6 +88,7 @@ drake_py_unittest(
     ],
     deps = [
         ":visualization",
+        "//bindings/pydrake/common/test_utilities:deprecation_py",
     ],
 )
 
@@ -93,6 +96,7 @@ drake_py_unittest(
     name = "plotting_test",
     deps = [
         ":visualization",
+        "//bindings/pydrake/common/test_utilities:deprecation_py",
     ],
 )
