+++ bindings/pydrake/visualization/_meldis.py
@@ -1,32 +1,7 @@
-"""
-MeshCat LCM Display Server (MeLDiS)
-
-A standalone program that can display Drake visualizations in MeshCat
-by listing for LCM messages that are broadcast by the simulation.
-
-This can stand in for the legacy ``drake-visualizer`` application of
-days past.
-
-From a Drake source build, run this as::
-
-  bazel run //tools:meldis &
-
-From a Drake binary release (including pip releases), run this as::
-
-  python3 -m pydrake.visualization.meldis
-
-In many cases, passing ``-w`` (i.e., ``--open-window``) to the program will be
-convenient::
-
-  bazel run //tools:meldis -- -w &
-"""
-
-import argparse
 import logging
 import numpy as np
 import sys
 import time
-import webbrowser
 
 from drake import (
     lcmt_contact_results_for_viz,
@@ -381,8 +356,8 @@ class Meldis:
     """
     MeshCat LCM Display Server (MeLDiS)
 
-    Offers a MeshCat vizualization server that listens for and draws Drake's
-    legacy LCM vizualization messages.
+    Offers a MeshCat visualization server that listens for and draws Drake's
+    legacy LCM visualization messages.
 
     If the meshcat_host parameter is not supplied, 'localhost' will be used by
     default.
@@ -391,7 +366,7 @@ class Meldis:
     """
 
     def __init__(self, *, meshcat_host=None, meshcat_port=None):
-        # Bookkeeping for update throtting.
+        # Bookkeeping for update throttling.
         self._last_update_time = time.time()
 
         # Bookkeeping for subscriptions, keyed by LCM channel name.
@@ -474,7 +449,7 @@ class Meldis:
         # However, if the sender is transmitting visualization messages at
         # a high rate (e.g., if a sim is running much faster than realtime),
         # then we should only pass some of them along to MeshCat to avoid
-        # flooding it. The hander merely records the message data; we'll
+        # flooding it. The handler merely records the message data; we'll
         # pass it along to MeshCat using our `self._should_update()` timer.
         def _on_message(data):
             self._message_pending_data[channel] = data
@@ -551,44 +526,3 @@ class Meldis:
             _logger.info("Meldis is exiting now; no browser was connected for"
                          f" >{idle_timeout} seconds")
             sys.exit(1)
-
-
-def _main():
-    configure_logging()
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--host", action="store",
-        help="The http listen host for MeshCat. If none is given, 'localhost'"
-        " will be used by default. In any case, the result will be printed to"
-        " the console.")
-    parser.add_argument(
-        "-p", "--port", action="store", metavar="NUM", type=int,
-        help="The http listen port for MeshCat. If none is given, a default"
-        " will be chosen and printed to the console.")
-    parser.add_argument(
-        "-t", "--open-tab", dest="browser_new",
-        action="store_const", const=2, default=None,
-        help="Open the MeshCat display in a browser tab.")
-    parser.add_argument(
-        "-w", "--open-window", dest="browser_new",
-        action="store_const", const=1, default=None,
-        help="Open the MeshCat display in a new browser window.")
-    parser.add_argument(
-        "--idle-timeout", metavar="TIME", type=float, default=15*60,
-        help="When no web browser has been connected for this many seconds,"
-        " this program will automatically exit. Set to 0 to run indefinitely.")
-    args = parser.parse_args()
-    meldis = Meldis(meshcat_host=args.host, meshcat_port=args.port)
-    if args.browser_new is not None:
-        url = meldis.meshcat.web_url()
-        webbrowser.open(url=url, new=args.browser_new)
-    idle_timeout = args.idle_timeout
-    if idle_timeout == 0.0:
-        idle_timeout = None
-    elif idle_timeout < 0.0:
-        parser.error("The --idle_timeout cannot be negative.")
-    meldis.serve_forever(idle_timeout=idle_timeout)
-
-
-if __name__ == "__main__":
-    _main()
