+++ bindings/pydrake/visualization/_visualization_extra.py
@@ -1,0 +1,9 @@
+from ._meldis import (
+    Meldis,
+)
+from ._plotting import (
+    plot_sublevelset_expression,
+    plot_sublevelset_quadratic,
+)
+
+__all__ = [x for x in globals() if not x.startswith("_")]
