--- bindings/pydrake/visualization/all.py
@@ -1,3 +1,0 @@
-from pydrake.visualization import *
-from .meldis import *
-from .plotting import *
