+++ bindings/pydrake/visualization/visualization_py.cc
@@ -16,6 +16,9 @@ Bindings for Visualization.
 
   // The order of these calls matters. Some modules rely on prior definitions.
   internal::DefineVisualizationConfig(m);
+
+  py::module::import("pydrake.visualization.meldis");
+  ExecuteExtraPythonCode(m, true);
 }
 
 }  // namespace pydrake
